# api/openapi.yaml
openapi: 3.1.0
info:
  title: Reticora CMDB API
  version: 1.0.0
  description: |
    Multi-tenant Configuration Management Database API for MSPs and ISPs.
    Endpoints require tenant-scoped authentication unless explicitly marked otherwise.
  contact:
    name: DataHub Chiemgau
  license:
    name: Proprietary
servers:
  - url: http://localhost:8080
    description: Local development
security:
  - BearerAuth: []
  - ApiKeyAuth: []
...
paths:
...
  /api/v1/discovery/review-items:
    get:
      operationId: listReviewItems
      tags: [Discovery]
      summary: List reconciliation review items
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
        - name: status
          in: query
          schema:
            type: string
            enum: [open, resolved, dismissed]
        - name: kind
          in: query
          schema:
            type: string
            enum: [ambiguous_identity, conflicting_values, unclassified_device]
      responses:
        '200':
          description: Paginated list of review items
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReviewItemListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /api/v1/iga/connectors:
    get:
      operationId: listIGAConnectors
      tags: [IGA]
      summary: List IGA connectors
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/IGAListResponse' } } } }
        '201': { description: Created, content: { application/json: { schema: { $ref: '#/components/schemas/IGAListResponse' } } } }
        '202': { description: Accepted, content: { application/json: { schema: { $ref: '#/components/schemas/IGAListResponse' } } } }
        '204': { description: Deleted }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalServerError' }
    post:
      operationId: createIGAConnector
      tags: [IGA]
      summary: Create an IGA connector
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/IGAObject' }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/IGAObject' } } } }
        '201': { description: Created, content: { application/json: { schema: { $ref: '#/components/schemas/IGAObject' } } } }
        '202': { description: Accepted, content: { application/json: { schema: { $ref: '#/components/schemas/IGAObject' } } } }
        '204': { description: Deleted }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalServerError' }
  /api/v1/iga/connectors/{id}:
    parameters:
        - $ref: '#/components/parameters/ResourceID'
    get:
      operationId: getIGAConnector
      tags: [IGA]
      summary: Get an IGA connector
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/IGAObject' } } } }
        '201': { description: Created, content: { application/json: { schema: { $ref: '#/components/schemas/IGAObject' } } } }
        '202': { description: Accepted, content: { application/json: { schema: { $ref: '#/components/schemas/IGAObject' } } } }
        '204': { description: Deleted }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalServerError' }
    patch:
      operationId: updateIGAConnector
      tags: [IGA]
      summary: Update an IGA connector
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/IGAObject' }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/IGAObject' } } } }
        '201': { description: Created, content: { application/json: { schema: { $ref: '#/components/schemas/IGAObject' } } } }
        '202': { description: Accepted, content: { application/json: { schema: { $ref: '#/components/schemas/IGAObject' } } } }
        '204': { description: Deleted }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalServerError' }
    delete:
      operationId: deleteIGAConnector
      tags: [IGA]
      summary: Delete an IGA connector
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/IGAObject' } } } }
        '201': { description: Created, content: { application/json: { schema: { $ref: '#/components/schemas/IGAObject' } } } }
        '202': { description: Accepted, content: { application/json: { schema: { $ref: '#/components/schemas/IGAObject' } } } }
        '204': { description: Deleted }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalServerError' }
  /api/v1/iga/connectors/{id}/test:
    parameters:
        - $ref: '#/components/parameters/ResourceID'
    post:
      operationId: testIGAConnector
      tags: [IGA]
      summary: Test an IGA connector
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/IGAObject' } } } }
        '201': { description: Created, content: { application/json: { schema: { $ref: '#/components/schemas/IGAObject' } } } }
        '202': { description: Accepted, content: { application/json: { schema: { $ref: '#/components/schemas/IGAObject' } } } }
        '204': { description: Deleted }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalServerError' }
  /api/v1/iga/connectors/{id}/sync:
    parameters:
        - $ref: '#/components/parameters/ResourceID'
    post:
      operationId: syncIGAConnector
      tags: [IGA]
      summary: Start connector reconciliation
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/IGAObject' } } } }
        '201': { description: Created, content: { application/json: { schema: { $ref: '#/components/schemas/IGAObject' } } } }
        '202': { description: Accepted, content: { application/json: { schema: { $ref: '#/components/schemas/IGAObject' } } } }
        '204': { description: Deleted }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalServerError' }
  /api/v1/iga/tasks:
    get:
      operationId: listIGAProvisioningTasks
      tags: [IGA]
      summary: List IGA provisioning tasks
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/IGAListResponse' } } } }
        '201': { description: Created, content: { application/json: { schema: { $ref: '#/components/schemas/IGAListResponse' } } } }
        '202': { description: Accepted, content: { application/json: { schema: { $ref: '#/components/schemas/IGAListResponse' } } } }
        '204': { description: Deleted }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalServerError' }
  /api/v1/iga/tasks/{id}:
    parameters:
        - $ref: '#/components/parameters/ResourceID'
    get:
      operationId: getIGAProvisioningTask
      tags: [IGA]
      summary: Get an IGA provisioning task
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/IGAObject' } } } }
        '201': { description: Created, content: { application/json: { schema: { $ref: '#/components/schemas/IGAObject' } } } }
        '202': { description: Accepted, content: { application/json: { schema: { $ref: '#/components/schemas/IGAObject' } } } }
        '204': { description: Deleted }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalServerError' }
  /api/v1/iga/tasks/{id}/retry:
    parameters:
        - $ref: '#/components/parameters/ResourceID'
    post:
      operationId: retryIGAProvisioningTask
      tags: [IGA]
      summary: Retry an IGA provisioning task
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/IGAObject' } } } }
        '201': { description: Created, content: { application/json: { schema: { $ref: '#/components/schemas/IGAObject' } } } }
        '202': { description: Accepted, content: { application/json: { schema: { $ref: '#/components/schemas/IGAObject' } } } }
        '204': { description: Deleted }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalServerError' }
  /api/v1/iga/jml-policies:
    get:
      operationId: listIGAJMLPolicies
      tags: [IGA]
      summary: List joiner/mover/leaver policies
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/IGAListResponse' } } } }
        '201': { description: Created, content: { application/json: { schema: { $ref: '#/components/schemas/IGAListResponse' } } } }
        '202': { description: Accepted, content: { application/json: { schema: { $ref: '#/components/schemas/IGAListResponse' } } } }
        '204': { description: Deleted }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalServerError' }
    post:
      operationId: createIGAJMLPolicy
      tags: [IGA]
      summary: Create a joiner/mover/leaver policy
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/IGAObject' }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/IGAObject' } } } }
        '201': { description: Created, content: { application/json: { schema: { $ref: '#/components/schemas/IGAObject' } } } }
        '202': { description: Accepted, content: { application/json: { schema: { $ref: '#/components/schemas/IGAObject' } } } }
        '204': { description: Deleted }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalServerError' }
  /api/v1/iga/jml/resolve:
    post:
      operationId: resolveIGALifecycle
      tags: [IGA]
      summary: Resolve an identity lifecycle change
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/IGAObject' }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/IGAArray' } } } }
        '201': { description: Created, content: { application/json: { schema: { $ref: '#/components/schemas/IGAArray' } } } }
        '202': { description: Accepted, content: { application/json: { schema: { $ref: '#/components/schemas/IGAArray' } } } }
        '204': { description: Deleted }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalServerError' }
  /api/v1/iga/access-requests:
    get:
      operationId: listIGAAccessRequests
      tags: [IGA]
      summary: List IGA access requests
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/IGAListResponse' } } } }
        '201': { description: Created, content: { application/json: { schema: { $ref: '#/components/schemas/IGAListResponse' } } } }
        '202': { description: Accepted, content: { application/json: { schema: { $ref: '#/components/schemas/IGAListResponse' } } } }
        '204': { description: Deleted }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalServerError' }
    post:
      operationId: createIGAAccessRequest
      tags: [IGA]
      summary: Create an IGA access request
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/IGAObject' }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/IGAObject' } } } }
        '201': { description: Created, content: { application/json: { schema: { $ref: '#/components/schemas/IGAObject' } } } }
        '202': { description: Accepted, content: { application/json: { schema: { $ref: '#/components/schemas/IGAObject' } } } }
        '204': { description: Deleted }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalServerError' }
  /api/v1/iga/access-requests/{id}/approve:
    parameters:
        - $ref: '#/components/parameters/ResourceID'
    post:
      operationId: approveIGAAccessRequest
      tags: [IGA]
      summary: Approve an IGA access request
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/IGAObject' }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/IGAObject' } } } }
        '201': { description: Created, content: { application/json: { schema: { $ref: '#/components/schemas/IGAObject' } } } }
        '202': { description: Accepted, content: { application/json: { schema: { $ref: '#/components/schemas/IGAObject' } } } }
        '204': { description: Deleted }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalServerError' }
  /api/v1/iga/access-requests/{id}/reject:
    parameters:
        - $ref: '#/components/parameters/ResourceID'
    post:
      operationId: rejectIGAAccessRequest
      tags: [IGA]
      summary: Reject an IGA access request
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/IGAObject' }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/IGAObject' } } } }
        '201': { description: Created, content: { application/json: { schema: { $ref: '#/components/schemas/IGAObject' } } } }
        '202': { description: Accepted, content: { application/json: { schema: { $ref: '#/components/schemas/IGAObject' } } } }
        '204': { description: Deleted }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalServerError' }
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  parameters:
...
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 50
      description: Maximum number of items to return.
    Offset:
      name: offset
      in: query
      schema:
        type: integer
        minimum: 0
        default: 0
      description: |
        Number of items to skip before returning data. Ignored when `cursor` is
        supplied.
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
    Unauthorized:
      description: Unauthorized
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
...
    InternalServerError:
      description: Internal server error
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
  schemas:
...
    IGAListResponse:
      type: object
      properties:
        data:
          type: array
          items: { $ref: '#/components/schemas/IGAObject' }
        total: { type: integer }
        limit: { type: integer }
        offset: { type: integer }
        has_more: { type: boolean }
...
    ReviewItemListResponse:
      allOf:
        - $ref: '#/components/schemas/PaginationEnvelope'
        - type: object
          required: [data]
          properties:
            data:
              type: array
              items:
                $ref: '#/components/schemas/ReviewItem'
...